src/stream_processing/flink_processing.py
from pyflink.datastream import StreamExecutionEnvironment
from pyflink.datastream.functions import MapFunction
from kafka import KafkaConsumer
import json

class MaintenanceMapFunction(MapFunction):
    def map(self, value):
        data = json.loads(value)
        # Implement logic to parse and process Kafka stream data
        # Example: return parsed data based on some condition
        if data["status"] == "normal":
            return f"Normal operation: {data}"
        else:
            return f"Alert: {data}"

if __name__ == "__main__":
    env = StreamExecutionEnvironment.get_execution_environment()
    data_stream = env.from_elements('{"status": "normal", "sensor_1": 40}', '{"status": "anomaly", "sensor_1": 70}')
    
    data_stream.map(MaintenanceMapFunction()).print()
    
    env.execute("Real-Time Maintenance Stream Processing")
