requirements.txt
kafka-python==2.0.2
pyspark==3.4.1
torch==2.0.1
minio==7.1.12
flink==1.16.0
seaborn==0.12.2
pandas==2.0.3
matplotlib==3.7.1